Clean up front end code + cleaned up path files
