+++ app/src/admin/DataGrid/noResultsOverlay.tsx
@@ -2,12 +2,16 @@ import { Typography } from '@mui/material'
 import { GridOverlay } from '@mui/x-data-grid'
 import { WebAsset } from '@mui/icons-material'
 
-export default function NoDishesOverlay() {
+interface Props {
+    value: string
+}
+
+export default function NoResultsOverlay({ value }: Props) {
     return (
         <GridOverlay style={{ flexDirection: 'column', marginTop: 4, marginBottom: 2 }}>
             <WebAsset color="action" fontSize="large" />
             <Typography mt={1} variant="overline">
-                No dishes
+                No {value}
             </Typography>
         </GridOverlay>
     )
