+++ app/src/routes/admin.tsx
@@ -44,6 +44,8 @@ import '../styles/admin.css'
 import leaf_white from '../assets/leaf-white.svg'
 import axios from 'axios'
 import { useAuth } from '../contexts/AuthContext'
+import { Box } from '@mui/material'
+import AdminDishes from '../admin/Dishes/dishesPage'
 
 function dishStatus(dishNumbers: number[]) {
     return (
@@ -398,7 +400,7 @@ function sortRows(rows) {
     return tableRows
 }
 
-function Admin() {
+function Admin({ path }: { path: string }) {
     const { currentUser, sessionToken } = useAuth()
     const [dishesUsed, setDishesUsed] = useState<any[]>([])
     const [transactionsUsed, setTransactionsUsed] = useState<any[]>([])
@@ -514,48 +516,52 @@ function Admin() {
         <>
             {/* on mobile */}
             <MobileView>
-                <div>
+                <Box sx={{ textAlign: 'center', mt: 2 }}>
                     <h1>Admin Panel</h1>
-                </div>
-                <img src={leaf_icon} alt="" />
-                <h2>You're on mobile! Please go to desktop to view admin panel.</h2>
+
+                    <img src={leaf_icon} alt="" />
+                    <h2>You're on mobile! Please go to desktop to view admin panel.</h2>
+                </Box>
             </MobileView>
 
             {/* on desktop */}
             <BrowserView>
-                <div className="d-flex">
-                    <Toolbar />
-                    <div style={{ marginTop: '48px', marginLeft: '40px', marginRight: '40px', flexGrow: 1 }}>
-                        <p className="sub-header-2">Home</p>
-                        {bar}
-                        <p className="sub-header-2" style={{ marginTop: 40 }}>
-                            Recent transactions
-                        </p>
+                {path == 'dishes' && <AdminDishes />}
+                {path == '' && (
+                    <div className="d-flex">
+                        <Toolbar />
+                        <div style={{ marginTop: '48px', marginLeft: '40px', marginRight: '40px', flexGrow: 1 }}>
+                            <p className="sub-header-2">Home</p>
+                            {bar}
+                            <p className="sub-header-2" style={{ marginTop: 40 }}>
+                                Recent transactions
+                            </p>
 
-                        <div className="d-flex" style={{ marginBottom: '16px' }}>
-                            {/* search Bar */}
+                            <div className="d-flex" style={{ marginBottom: '16px' }}>
+                                {/* search Bar */}
 
-                            <input
-                                className="search-container d-flex"
-                                type="text"
-                                placeholder="Type text here..."
-                                onChange={searchChange}
-                                value={searchInput}
-                                style={{ marginRight: '8px' }}
-                            />
+                                <input
+                                    className="search-container d-flex"
+                                    type="text"
+                                    placeholder="Type text here..."
+                                    onChange={searchChange}
+                                    value={searchInput}
+                                    style={{ marginRight: '8px' }}
+                                />
 
-                            <button className="search-b d-flex" onClick={handleClick}>
-                                <p className="sub-header-3">Search</p>
-                            </button>
+                                <button className="search-b d-flex" onClick={handleClick}>
+                                    <p className="sub-header-3">Search</p>
+                                </button>
 
-                            <div className="d-flex justify-content-end" style={{ width: '100%' }}>
-                                {add}
+                                <div className="d-flex justify-content-end" style={{ width: '100%' }}>
+                                    {add}
+                                </div>
                             </div>
+                            {table}
+                            {row}
                         </div>
-                        {table}
-                        {row}
                     </div>
-                </div>
+                )}
             </BrowserView>
         </>
     )
